everything in double

+++ DynamicArrays/DynamicArrays/CountDifferent.cpp
@@ -1,5 +1,5 @@
 #include "SecondTasks.H"
-int32_t CountDifferentNumbers(int32_t* a, int32_t size)
+int32_t CountDifferentNumbers(double* a, int32_t size)
 {
 	int32_t ammount{1};
 	for (int32_t i{}; i < size - 1; ++i)
